refactor: split Login and Project Sheet scenes's protocols into separated folder and files
