+++ Pharus/Scenes/Login/Protocols/LoginFlow.swift
@@ -1,0 +1,12 @@
+//
+//  LoginFlow.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol LoginFlow {
+    func showHome(student: StudentModel)
+}
