+++ Pharus/Scenes/Login/Protocols/LoginPresenterProtocol.swift
@@ -1,0 +1,14 @@
+//
+//  LoginPresenterProtocol.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol LoginPresenterProtocol {
+    func isValidEmail(email: String) -> Bool
+    func isValidPassword(password: String) -> Bool
+    func loginUser(email: String, password: String) -> Bool
+}
