+++ Pharus/Scenes/Login/Protocols/LoginViewDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  LoginViewDelegate.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol LoginViewDelegate: AnyObject {
+    func loginButtonPressed()
+}
