+++ Pharus/Scenes/Project Sheet/Protocols/ProjectSheetDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  ProjectSheetDelegate.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol ProjectSheetDelegate: AnyObject {
+    func closeButtonTapped()
+}
