+++ Pharus/Scenes/Project Sheet/Protocols/ProjectSheetFlow.swift
@@ -1,0 +1,12 @@
+//
+//  ProjectSheetFlow.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol ProjectSheetFlow {
+    func popView()
+}
