+++ Pharus/Scenes/Project Sheet/Protocols/ProjectSheetProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  ProjectSheetProtocol.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol ProjectSheetProtocol {
+    func popView()
+}
